requirements.txt
asgiref==3.11.1
cfgv==3.5.0
distlib==0.4.0
dj-database-url==3.1.0
Django==5.2.11
django-stubs==5.2.9
django-stubs-ext==5.2.9
filelock==3.21.2
gunicorn==25.1.0
identify==2.6.16
librt==0.8.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.7.1
pre_commit==4.5.1
psycopg2-binary==2.9.11
python-decouple==3.8
PyYAML==6.0.3
sqlparse==0.5.5
tomli==2.4.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
virtualenv==20.36.1
whitenoise==6.11.0
